Only change working directory to directory when asked.
If the user wants to change to <dir> before executing <cmd>,
they now need to 'stampaldf --cd <dir> <cmd>', or just
( cd dir && stampaldf . <cmd> ).

The reason to not do this by default is that if 'cmd' has
any references to filesystem paths that are relative to
the callers working directory, then they'll be broken.

Ie, previously this would fail:

    mkdir foo
    touch bar
    touch foo/subfile
    stampaldf foo rm bar

Which is weird.  You don't expect commands to change their
working directory.

If I wanted that, I could just do this:

    sh -c 'cd foo && stampaldf . rm subfile'

Or, alternatively if you want it you can now do:

    stampaldf --cd . rm subfile